Rename fixtures/paypal-response to fixtures/paypal-create-order-response
