+++ package.json
@@ -45,7 +45,7 @@
   },
   "dependencies": {
     "@shiftcommerce/shift-node-api": "^0.1.1",
-    "@shiftcommerce/shift-react-components": "shiftcommerce/shift-react-components#feature/790_checkout_integration",
+    "@shiftcommerce/shift-react-components": "shiftcommerce/shift-react-components#issue/790_updates",
     "axios": "^0.18.0",
     "js-cookie": "^2.2.0",
     "qs": "^6.6.0",
