+++ test/pages/payment-method.test.js
@@ -9,7 +9,7 @@ import * as CheckoutActions from '../../src/actions/checkout-actions'
 import * as CartActions from '../../src/actions/cart-actions'
 
 // Fixtures
-const payPalResponse = require('../fixtures/paypal-response')
+const payPalResponse = require('../fixtures/paypal-create-order-response')
 
 jest.mock('next/config', () => () => ({
   publicRuntimeConfig: {}
